添加item id

+++ Daisy/src/tv/ismar/daisy/data/http/newvip/paylayer/Package.java
@@ -6,7 +6,7 @@ package tv.ismar.daisy.data.http.newvip.paylayer;
 public class Package {
     private String title;
 
-    private int price;
+    private float price;
 
     private String duration;
 
@@ -14,34 +14,43 @@ public class Package {
 
     private int package_pk;
 
-    public void setTitle(String title){
+    public void setTitle(String title) {
         this.title = title;
     }
-    public String getTitle(){
+
+    public String getTitle() {
         return this.title;
     }
-    public void setPrice(int price){
+
+    public void setPrice(int price) {
         this.price = price;
     }
-    public int getPrice(){
+
+    public float getPrice() {
         return this.price;
     }
-    public void setDuration(String duration){
+
+    public void setDuration(String duration) {
         this.duration = duration;
     }
-    public String getDuration(){
+
+    public String getDuration() {
         return this.duration;
     }
-    public void setVertical_url(String vertical_url){
+
+    public void setVertical_url(String vertical_url) {
         this.vertical_url = vertical_url;
     }
-    public String getVertical_url(){
+
+    public String getVertical_url() {
         return this.vertical_url;
     }
-    public void setPackage_pk(int package_pk){
+
+    public void setPackage_pk(int package_pk) {
         this.package_pk = package_pk;
     }
-    public int getPackage_pk(){
+
+    public int getPackage_pk() {
         return this.package_pk;
     }
